ビルドバリアントに応じて，注入するモジュールを変更するようにした

+++ app/src/main/java/com/example/hmiyado/sampo/SampoApplication.kt
@@ -7,13 +7,17 @@ import com.example.hmiyado.sampo.repository.compass.CompassServiceImpl
 import com.example.hmiyado.sampo.repository.compass.CompassServiceVirtualImpl
 import com.example.hmiyado.sampo.repository.location.*
 import com.example.hmiyado.sampo.service.LocationAndroidService
-import com.github.salomonbrys.kodein.*
+import com.github.salomonbrys.kodein.Kodein
+import com.github.salomonbrys.kodein.KodeinAware
 import com.github.salomonbrys.kodein.android.activityScope
 import com.github.salomonbrys.kodein.android.androidModule
 import com.github.salomonbrys.kodein.android.autoActivitySingleton
+import com.github.salomonbrys.kodein.lazy
+import com.github.salomonbrys.kodein.singleton
 import com.jakewharton.threetenabp.AndroidThreeTen
 import io.realm.Realm
 import io.realm.RealmConfiguration
+import org.jetbrains.anko.locationManager
 import org.jetbrains.anko.sensorManager
 import timber.log.Timber
 
@@ -26,11 +30,14 @@ class SampoApplication : Application(), KodeinAware {
     override val kodein by Kodein.lazy {
         import(androidModule)
 
-        bind<LocationService>() with singleton { LocationServiceVirtualImpl() }
-        bind<LocationService>("real") with singleton { LocationServiceImpl(instance()) }
+        if (BuildConfig.FLAVOR == "mock") {
+            bind<LocationService>() with singleton { LocationServiceVirtualImpl() }
+            bind<CompassService>() with singleton { CompassServiceVirtualImpl() }
+        } else if (BuildConfig.FLAVOR == "real") {
+            bind<LocationService>() with singleton { LocationServiceImpl(locationManager) }
+            bind<CompassService>() with singleton { CompassServiceImpl(sensorManager) }
+        }
         bind<LocationRepository>() with autoActivitySingleton { LocationRepositoryRealmImpl() }
-        bind<CompassService>("real") with singleton { CompassServiceImpl(sensorManager) }
-        bind<CompassService>() with singleton { CompassServiceVirtualImpl() }
     }
 
 
@@ -40,10 +47,12 @@ class SampoApplication : Application(), KodeinAware {
         registerActivityLifecycleCallbacks(activityScope.lifecycleManager)
         Realm.init(baseContext)
         val realmConfig = RealmConfiguration.Builder()
-                .name("Location.realm")
+                .name(BuildConfig.BUILD_TYPE + BuildConfig.FLAVOR)
                 .schemaVersion(1)
                 .build()
-        Realm.deleteRealm(realmConfig)
+        if (!BuildConfig.DEBUG) {
+            Realm.deleteRealm(realmConfig)
+        }
         Realm.setDefaultConfiguration(realmConfig)
 
         Timber.plant(Timber.DebugTree())
